Resolve various encoding + type errors with fromJson and toJson
All this work so that way representations from both the database and from redis can be handled from one system. However, this is not the most type secure system based on the way I implemented it with the ternary operators. Other than the lack of type safety with ternary operators, I believe the rest is relatively solid.